Update dagger2 and improve build file
Updates daggers2 to a later version, reverts change that adds a
dependency on Guava v19 as AOSP currently only has v18. See
README.android for more details.

Made a couple of changes to the build to make it reusable.

1) Added dagger2_annotation_processor.mk to define values that
are needed by modules that depend on dagger2.
2) Added java_annotation_processor.mk to do the magic needed to
make the annotation processors work properly.

Change-Id: Ie0e8b0e34c961f64ccd0a45bd93a7ea6a216ddf3