Move host dependencies to Android.bp
Also move dagger2-inject-host to a static dependency, all users of
dagger2-host include dagger2-inject-host statically, and depending
on as non-static library causes it to be installed unnecessarily.

See build/soong/README.md for more information.

The libraries in lib/ are used outside the project by host tools
that are prerequisites for converting device modules to soong.

Test: m -j checkbuild
Change-Id: I3f8164c3a67b37f88fca888cabac290d2c927d1e